+++ trafaret/contrib/object_id.py
@@ -33,11 +33,11 @@ class MongoId(Trafaret):
 
     def check_and_return(self, value):
         if not self.allow_blank and value is None:
-            self._failure("blank value is not allowed")
+            self._failure("blank value is not allowed", code='empty_value')
         if isinstance(value, self.convertable) or value is None:
             try:
                 return ObjectId(value)
             except InvalidId as e:
-                self._failure(str(e))
+                self._failure(str(e), code='invalid_objectid')
 
-        self._failure('value is not %s' % self.value_type.__name__)
+        self._failure('value is not %s' % self.value_type.__name__, code='not_objectid')