+++ docs/changelog.rst
@@ -9,6 +9,7 @@ Changelog
 – trafarets are not mutable
 – DataError has new ``code`` attribute, self.failure has ``code`` argument
 - OnError has ``code`` argument too
+- New `DataError.to_struct` method that returns errors in more consistent way
 – String, AnyString, Bytes
 – Int, ToInt, Float, ToFloat
 - ToDecimal